Fixed nasty type-o causing issues on 64-bit operations on all non win32 platforms

+++ lib/mcdefs.h
@@ -36,8 +36,8 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <windows.h>
 #else
 #include <sys/types.h>
-typedef u_int32_t UINT64;
-typedef u_int32_t UINT8;
+typedef u_int64_t UINT64;
+typedef u_int8_t UINT8;
 #endif
 
 #endif
